refactor: move utils and build iokit package
