+++ pyproject.toml
@@ -1,10 +1,10 @@
 [project]
 name = "iokit"
+version = "0.3.5"
 description = "Input Output Kit"
 readme = "README.md"
 requires-python = ">=3.10"
 license = {text = "MIT"}
-dynamic = ["version"]
 authors = [
     {name = "Vladislav A. Proskurov", email = "rilshok@pm.me"},
 ]
@@ -26,10 +26,6 @@ dependencies = [
 ]
 
 
-[tool.setuptools.dynamic]
-version = {attr = "iokit.__version__"}
-
-
 [project.optional-dependencies]
 dev = ["iokit[lint,test]"]
 lint = [
@@ -45,6 +41,11 @@ test = [
     "pytest-cov==6.0.0",
     "pytest-xdist>=3.6.1",
 ]
+
+
+[build-system]
+requires = ["uv_build>=0.11.29,<0.12.0"]
+build-backend = "uv_build"
 
 
 [project.urls]
