+++ src/iokit/extensions/json.py
@@ -6,7 +6,7 @@ from datetime import datetime
 from functools import lru_cache
 from typing import Any
 
-from iokit.state import State, StateName
+from iokit.utils.state import State, StateName
 
 
 @lru_cache
