+++ src/iokit/extensions/yaml.py
@@ -4,7 +4,7 @@ from datetime import datetime
 
 import yaml
 
-from iokit.state import State, StateName
+from iokit.utils.state import State, StateName
 
 
 class Yaml(State, suffix="yaml"):
